+++ src/dftbp/dftbplus/main.F90
@@ -58,9 +58,10 @@ module dftbp_dftbplus_main
   use dftbp_dftb_spinorbit, only : addOnsiteSpinOrbitHam, getOnsiteSpinOrbitEnergy
   use dftbp_dftb_stress, only : getBlockiStress, getBlockStress, getkineticstress, getNonSCCStress
   use dftbp_dftb_thirdorder, only : TThirdOrder
+  use dftbp_dftbplus_apicallback, only : TAPICallback
   use dftbp_dftbplus_eigenvects, only : diagDenseMtx
   use dftbp_dftbplus_forcetypes, only : forceTypes
-  use dftbp_dftbplus_initprogram, only : TDftbPlusMain, TNegfInt
+  use dftbp_dftbplus_initprogram, only : TDangerousChange, TDftbPlusMain, TNegfInt
   use dftbp_dftbplus_inputdata, only : TNEGFInfo
   use dftbp_dftbplus_mainio, only : openOutputFile, printBlankLine, printElecConstrHeader,&
       & printElecConstrInfo, printEnergies, printForceNorm, printGeostepInfo,&
@@ -1299,8 +1300,9 @@ contains
             & this%iSparseStart, this%img2CentCell, this%orb, this%species, this%coord,&
             & this%tPeriodic, this%tHelical, this%eigvecsReal, this%parallelKS, this%rhoPrim,&
             & this%SSqrReal, this%rhoSqrReal, this%q0, this%densityMatrix, this%hybridXc,&
-            & this%reks, errstatus)
+            & this%reks, this%apiCallBack, errstatus)
         @:PROPAGATE_ERROR(errStatus)
+
         call getMullikenPopulationL(env, this%denseDesc, this%neighbourList, this%nNeighbourSK,&
             & this%img2CentCell, this%iSparseStart, this%orb, this%rhoPrim, this%ints,&
             & this%iRhoPrim, this%qBlockOut, this%qiBlockOut, this%qNetAtom, this%reks)
@@ -1327,7 +1329,7 @@ contains
             & this%neighbourList, this%nNeighbourSK, this%iSparseStart, this%img2CentCell,&
             & this%orb, this%species, this%coord, this%tPeriodic, this%tHelical, this%eigVecsReal,&
             & this%parallelKS, this%densityMatrix, this%rhoPrim, this%SSqrReal, this%rhoSqrReal,&
-            & this%hybridXc, errStatus)
+            & this%hybridXc, this%apiCallBack, errStatus)
         @:PROPAGATE_ERROR(errStatus)
         ! For hybrid xc-functional calculations deduct atomic charges from deltaRho
         if (this%isHybridXc) then
@@ -1452,7 +1454,8 @@ contains
               & this%hybridXc, this%eigen, this%filling, this%rhoPrim, this%xi, this%orbitalL,&
               & this%HSqrReal, this%SSqrReal, this%eigvecsReal, this%iRhoPrim, this%HSqrCplx,&
               & this%SSqrCplx, this%eigvecsCplx, this%rhoSqrReal, this%densityMatrix,&
-              & this%nNeighbourCam, this%nNeighbourCamSym, this%deltaDftb, errStatus)
+              & this%nNeighbourCam, this%nNeighbourCamSym, this%deltaDftb, this%apiCallBack,&
+              & this%dangerousChanges, errStatus)
           if (errStatus%hasError()) call error(errStatus%message)
 
           if (this%tWriteBandDat) then
@@ -2742,7 +2745,9 @@ contains
       & tRealHS, tSpinSharedEf, tSpinOrbit, tDualSpinOrbit, tFillKSep, tFixEf, tMulliken,&
       & iDistribFn, tempElec, nEl, parallelKS, Ef, mu, energy, hybridXc, eigen, filling, rhoPrim,&
       & xi, orbitalL, HSqrReal, SSqrReal, eigvecsReal, iRhoPrim, HSqrCplx, SSqrCplx, eigvecsCplx,&
-      & rhoSqrReal, densityMatrix, nNeighbourCam, nNeighbourCamSym, deltaDftb, errStatus)
+      & rhoSqrReal, densityMatrix, nNeighbourCam, nNeighbourCamSym, deltaDftb, apiCallBack,&
+      & dangerousChanges, errStatus)
+
     use dftbp_elecsolvers_dmsolvertypes, only : densityMatrixTypes
 
     !> Environment settings
@@ -2910,6 +2915,12 @@ contains
     !> Determinant derived type
     type(TDftbDeterminants), intent(inout) :: deltaDftb
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(TAPICallback), intent(inout), allocatable :: apiCallBack
+
+    !> Possibly fatal situations to check for at run-time
+    type(TDangerousChange) :: dangerousChanges
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
@@ -2932,7 +2943,7 @@ contains
           & iDistribFn, tempElec, nEl, parallelKS, Ef, energy, hybridXc, eigen, filling, rhoPrim,&
           & xi, orbitalL, HSqrReal, SSqrReal, eigvecsReal, iRhoPrim, HSqrCplx, SSqrCplx,&
           & eigvecsCplx, rhoSqrReal, densityMatrix, nNeighbourCam, nNeighbourCamSym, deltaDftb,&
-          & errStatus)
+          & apiCallBack, dangerousChanges, errStatus)
       @:PROPAGATE_ERROR(errStatus)
 
     case(densityMatrixTypes%elecSolverProvided)
@@ -2952,6 +2963,7 @@ contains
         call ud2qm(rhoPrim)
 
       else
+
         call electronicSolver%elsi%getDensity(env, denseDesc, ints%hamiltonian, ints%overlap,&
             & neighbourList, nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, kPoint,&
             & kWeight, tHelical, orb, species, coord, tRealHS, tSpinSharedEf, tSpinOrbit,&
@@ -2973,7 +2985,8 @@ contains
       & tRealHS, tSpinSharedEf, tSpinOrbit, tDualSpinOrbit, tFillKSep, tFixEf, tMulliken,&
       & iDistribFn, tempElec, nEl, parallelKS, Ef, energy, hybridXc, eigen, filling, rhoPrim, xi,&
       & orbitalL, HSqrReal, SSqrReal, eigvecsReal, iRhoPrim, HSqrCplx, SSqrCplx, eigvecsCplx,&
-      & rhoSqrReal, densityMatrix, nNeighbourCam, nNeighbourCamSym, deltaDftb, errStatus)
+      & rhoSqrReal, densityMatrix, nNeighbourCam, nNeighbourCamSym, deltaDftb, apiCallBack,&
+      & dangerousChanges, errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3131,6 +3144,12 @@ contains
     !> Determinant derived type
     type(TDftbDeterminants), intent(inout) :: deltaDftb
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(Tapicallback), intent(inout), allocatable :: apiCallBack
+
+    !> Possibly fatal situations to check for at run-time
+    type(TDangerousChange) :: dangerousChanges
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
@@ -3144,21 +3163,22 @@ contains
             & symNeighbourList, nNeighbourSK, iSparseStart, img2CentCell, orb, tPeriodic, tHelical,&
             & coord, electronicSolver, parallelKS, hybridXc, densityMatrix%deltaRhoIn,&
             & nNeighbourCam, nNeighbourCamSym, HSqrReal, SSqrReal, eigVecsReal, eigen(:,1,:),&
-            & errStatus)
+            & apiCallBack, dangerousChanges, errStatus)
         @:PROPAGATE_ERROR(errStatus)
       else
         call buildAndDiagDenseCplxHam(env, denseDesc, ints, kPoint, kWeight, neighbourList,&
             & symNeighbourList, nNeighbourSK, iSparseStart, img2CentCell, rCellVecs, iCellVec,&
             & recVecs2p, cellVec, electronicSolver, parallelKS, tHelical, orb, species, coord,&
             & hybridXc, densityMatrix, nNeighbourCamSym, HSqrCplx, SSqrCplx, eigVecsCplx, eigen,&
-            & errStatus)
+            & apiCallBack, dangerousChanges, errStatus)
         @:PROPAGATE_ERROR(errStatus)
       end if
     else
       call buildAndDiagDensePauliHam(env, denseDesc, ints, kPoint, neighbourList,&
           & nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, orb, electronicSolver,&
           & parallelKS, hybridXc, densityMatrix%deltaRhoInCplx, nNeighbourCam, nNeighbourCamSym,&
-          & HSqrCplx, SSqrCplx, eigVecsCplx, eigen(:,:,1), errStatus, xi, species)
+          & HSqrCplx, SSqrCplx, eigVecsCplx, eigen(:,:,1), apiCallBack, dangerousChanges,&
+          & errStatus, xi, species)
       @:PROPAGATE_ERROR(errStatus)
     end if
     call env%globalTimer%stopTimer(globalTimers%diagonalization)
@@ -3171,13 +3191,14 @@ contains
       if (tRealHS) then
         call getDensityFromRealEigvecs(env, denseDesc, filling(:,1,:), neighbourList, nNeighbourSK,&
             & iSparseStart, img2CentCell, orb, species, coord, tPeriodic, tHelical, eigVecsReal,&
-            & parallelKS, densityMatrix, rhoPrim, SSqrReal, rhoSqrReal, hybridXc, errStatus)
+            & parallelKS, densityMatrix, rhoPrim, SSqrReal, rhoSqrReal, hybridXc, apiCallBack,&
+            & errStatus)
         @:PROPAGATE_ERROR(errStatus)
       else
         call getDensityFromCplxEigvecs(env, denseDesc, filling, kPoint, kWeight, neighbourList,&
             & nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, orb,&
             & parallelKS, tHelical, species, coord, eigvecsCplx, densityMatrix, rhoPrim, SSqrCplx,&
-            & hybridXc, errStatus)
+            & hybridXc, apiCallBack, errStatus)
         @:PROPAGATE_ERROR(errStatus)
       end if
       call ud2qm(rhoPrim)
@@ -3187,7 +3208,8 @@ contains
       call getDensityFromPauliEigvecs(env, denseDesc, tRealHS, tSpinOrbit, tDualSpinOrbit,&
           & tMulliken, kPoint, kWeight, filling(:,:,1), neighbourList, nNeighbourSK, orb,&
           & iSparseStart, img2CentCell, iCellVec, cellVec, species, parallelKS, eigVecsCplx,&
-          & SSqrCplx, energy, densityMatrix, rhoPrim, xi, orbitalL, iRhoPrim, errStatus)
+          & SSqrCplx, energy, densityMatrix, rhoPrim, xi, orbitalL, iRhoPrim, apiCallBack,&
+          & errStatus)
       @:PROPAGATE_ERROR(errStatus)
       filling(:,:,1) = 0.5_dp * filling(:,:,1)
     end if
@@ -3200,7 +3222,8 @@ contains
   subroutine buildAndDiagDenseRealHam(env, denseDesc, ints, species, neighbourList,&
       & symNeighbourList, nNeighbourSK, iSparseStart, img2CentCell, orb, tPeriodic, tHelical,&
       & coord, electronicSolver, parallelKS, hybridXc, deltaRhoIn, nNeighbourCam,&
-      & nNeighbourCamSym, HSqrReal, SSqrReal, eigvecsReal, eigen, errStatus)
+      & nNeighbourCamSym, HSqrReal, SSqrReal, eigvecsReal, eigen, apiCallBack, dangerousChanges,&
+      & errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3271,10 +3294,17 @@ contains
     !> Eigenvalues
     real(dp), intent(out) :: eigen(:,:)
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(Tapicallback), intent(in), allocatable :: apiCallBack
+
+    !> Possibly fatal situations to check for at run-time
+    type(TDangerousChange) :: dangerousChanges
+
     !> Status of operation
     type(TStatus), intent(inout) :: errStatus
 
-    integer :: iKS, iSpin
+    integer :: iKS, iSpin, iK
+    logical :: isSChanged, isHChanged
 
     eigen(:,:) = 0.0_dp
 
@@ -3304,6 +3334,25 @@ contains
         @:PROPAGATE_ERROR(errStatus)
       end if
 
+      if (allocated(apiCallBack)) then
+        call apiCallBack%invokeHSCallBack(parallelKS%localKS(:,iKS),&
+            & electronicSolver%hasCholesky(iKS), SSqrReal, HSqrReal, isSChanged, isHChanged,&
+            & denseDesc)
+        if (isSChanged) then
+          if (dangerousChanges%overlap) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the overlap matrix, aborting")
+          end if
+          call warning("ASI interface has modified the overlap matrix at runtime")
+        end if
+        if (isHChanged) then
+          if (dangerousChanges%hamiltonian) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the hamiltonian matrix, aborting")
+          end if
+          call warning("ASI interface has modified the hamiltonian matrix at runtime")
+        end if
+
+      endif
+
       call diagDenseMtxBlacs(electronicSolver, 1, 'V', denseDesc%blacsOrbSqr, HSqrReal, SSqrReal,&
           & eigen(:,iSpin), eigvecsReal(:,:,iKS), errStatus)
       @:PROPAGATE_ERROR(errStatus)
@@ -3330,6 +3379,23 @@ contains
         @:PROPAGATE_ERROR(errStatus)
       end if
 
+      if (allocated(apiCallBack)) then
+        call apiCallBack%invokeHSCallBack(parallelKS%localKS(:,iKS),&
+            & electronicSolver%hasCholesky(iKS), SSqrReal, HSqrReal, isSChanged, isHChanged)
+        if (isSChanged) then
+          if (dangerousChanges%overlap) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the overlap matrix, aborting")
+          end if
+          call warning("ASI interface has modified the overlap matrix at runtime")
+        end if
+        if (isHChanged) then
+          if (dangerousChanges%hamiltonian) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the hamiltonian matrix, aborting")
+          end if
+          call warning("ASI interface has modified the hamiltonian matrix at runtime")
+        end if
+      end if
+
       ! Warning: SSqrReal gets overwritten here
       call diagDenseMtx(env, electronicSolver, 'V', HSqrReal, SSqrReal, eigen(:, iSpin),&
           & errStatus)
@@ -3350,7 +3416,8 @@ contains
   subroutine buildAndDiagDenseCplxHam(env, denseDesc, ints, kPoint, kWeight, neighbourList,&
       & symNeighbourList, nNeighbourSK, iSparseStart, img2CentCell, rCellVecs, iCellVec, recVecs2p,&
       & cellVec, electronicSolver, parallelKS, tHelical, orb, species, coord, hybridXc,&
-      & densityMatrix, nNeighbourCamSym, HSqrCplx, SSqrCplx, eigvecsCplx, eigen, errStatus)
+      & densityMatrix, nNeighbourCamSym, HSqrCplx, SSqrCplx, eigvecsCplx, eigen, apiCallBack,&
+      & dangerousChanges, errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3433,6 +3500,12 @@ contains
     !> Eigenvalues
     real(dp), intent(out) :: eigen(:,:,:)
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
+    type(Tapicallback), intent(in), allocatable :: apiCallBack
+
+    !> Possibly fatal situations to check for at run-time
+    type(TDangerousChange) :: dangerousChanges
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
@@ -3444,6 +3517,8 @@ contains
 
     !! Indices for k-points and spins + composite
     integer :: iK, iSpin, iKS
+
+    logical :: isSChanged, isHChanged
 
     eigen(:,:,:) = 0.0_dp
 
@@ -3513,9 +3588,28 @@ contains
         HSqrCplx(:,:) = HSqrCplx + HSqrCplxCam(:,:, densityMatrix%iKiSToiGlobalKS(iK, iSpin))
       end if
 
+      if (allocated(apiCallBack)) then
+        call apiCallBack%invokeHSCallBack(parallelKS%localKS(:,iKS),&
+            & electronicSolver%hasCholesky(iKS), SSqrCplx, HSqrCplx, isSChanged, isHChanged,&
+            & denseDesc)
+        if (isSChanged) then
+          if (dangerousChanges%overlap) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the overlap matrix, aborting")
+          end if
+          call warning("ASI interface has modified the overlap matrix at runtime")
+        end if
+        if (isHChanged) then
+          if (dangerousChanges%hamiltonian) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the hamiltonian matrix, aborting")
+          end if
+          call warning("ASI interface has modified the hamiltonian matrix at runtime")
+        end if
+      endif
+
       call diagDenseMtxBlacs(env, electronicSolver, iKS, 'V', denseDesc%blacsOrbSqr, HSqrCplx,&
           & SSqrCplx, eigen(:,iK,iSpin), eigvecsCplx(:,:,iKS), errStatus)
       @:PROPAGATE_ERROR(errStatus)
+
     #:else
       call env%globalTimer%startTimer(globalTimers%sparseToDense)
       if (tHelical) then
@@ -3539,6 +3633,23 @@ contains
         HSqrCplx(:,:) = HSqrCplx + HSqrCplxCam(:,:, densityMatrix%iKiSToiGlobalKS(iK, iSpin))
       end if
 
+      if (allocated(apiCallBack)) then
+        call apiCallBack%invokeHSCallBack(parallelKS%localKS(:,iKS),&
+            & electronicSolver%hasCholesky(iKS), SSqrCplx, HSqrCplx, isSChanged, isHChanged)
+        if (isSChanged) then
+          if (dangerousChanges%overlap) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the overlap matrix, aborting")
+          end if
+          call warning("ASI interface has modified the overlap matrix at runtime")
+        end if
+        if (isHChanged) then
+          if (dangerousChanges%hamiltonian) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the hamiltonian matrix, aborting")
+          end if
+          call warning("ASI interface has modified the hamiltonian matrix at runtime")
+        end if
+      endif
+
       call diagDenseMtx(env, electronicSolver, 'V', HSqrCplx, SSqrCplx, eigen(:, iK, iSpin),&
           & errStatus)
       @:PROPAGATE_ERROR(errStatus)
@@ -3558,7 +3669,7 @@ contains
   subroutine buildAndDiagDensePauliHam(env, denseDesc, ints, kPoint, neighbourList,&
       & nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, orb, electronicSolver,&
       & parallelKS, hybridXc, deltaRhoIn, nNeighbourCam, nNeighbourCamSym, HSqrCplx,&
-      & SSqrCplx, eigvecsCplx, eigen, errStatus, xi, species)
+      & SSqrCplx, eigvecsCplx, eigen, apiCallBack, dangerousChanges, errStatus, xi, species)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3623,6 +3734,12 @@ contains
     !> Eigenvalues (orbital, kpoint)
     real(dp), intent(out) :: eigen(:,:)
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(Tapicallback), intent(in), allocatable :: apiCallBack
+
+    !> Possibly fatal situations to check for at run-time
+    type(TDangerousChange) :: dangerousChanges
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
@@ -3633,6 +3750,7 @@ contains
     integer, intent(in), optional :: species(:)
 
     integer :: iKS, iK
+    logical :: isSChanged, isHChanged
 
     eigen(:,:) = 0.0_dp
     do iKS = 1, parallelKS%nLocalKS
@@ -3676,13 +3794,52 @@ contains
       end if
       call env%globalTimer%stopTimer(globalTimers%sparseToDense)
     #:if WITH_SCALAPACK
+
+      if (allocated(apiCallBack)) then
+        call apiCallBack%invokeHSCallBack(parallelKS%localKS(:,iKS),&
+            & electronicSolver%hasCholesky(iKS), SSqrCplx, HSqrCplx, isSChanged, isHChanged,&
+            & denseDesc)
+        if (isSChanged) then
+          if (dangerousChanges%overlap) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the overlap matrix, aborting")
+          end if
+          call warning("ASI interface has modified the overlap matrix at runtime")
+        end if
+        if (isHChanged) then
+          if (dangerousChanges%hamiltonian) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the hamiltonian matrix, aborting")
+          end if
+          call warning("ASI interface has modified the hamiltonian matrix at runtime")
+        end if
+      endif
+
       call diagDenseMtxBlacs(env, electronicSolver, iKS, 'V', denseDesc%blacsOrbSqr, HSqrCplx,&
           & SSqrCplx, eigen(:,iK), eigvecsCplx(:,:,iKS), errStatus)
       @:PROPAGATE_ERROR(errStatus)
+
     #:else
+
+      if (allocated(apiCallBack)) then
+        call apiCallBack%invokeHSCallBack(parallelKS%localKS(:,iKS),&
+            & electronicSolver%hasCholesky(iKS), SSqrCplx, HSqrCplx, isSChanged, isHChanged)
+        if (isSChanged) then
+          if (dangerousChanges%overlap) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the overlap matrix, aborting")
+          end if
+          call warning("ASI interface has modified the overlap matrix at runtime")
+        end if
+        if (isHChanged) then
+          if (dangerousChanges%hamiltonian) then
+            @:RAISE_ERROR(errStatus, -1, "ASI interface changed the hamiltonian matrix, aborting")
+          end if
+          call warning("ASI interface has modified the hamiltonian matrix at runtime")
+        end if
+      endif
+
       call diagDenseMtx(env, electronicSolver, 'V', HSqrCplx, SSqrCplx, eigen(:,iK), errStatus)
       @:PROPAGATE_ERROR(errStatus)
       eigvecsCplx(:,:,iKS) = HSqrCplx
+
     #:endif
     end do
 
@@ -3696,7 +3853,7 @@ contains
   !> Creates sparse density matrix from real eigenvectors.
   subroutine getDensityFromRealEigvecs(env, denseDesc, filling, neighbourList, nNeighbourSK,&
       & iSparseStart, img2CentCell, orb, species, coord, tPeriodic, tHelical, eigvecs, parallelKS,&
-      & densityMatrix, rhoPrim, work, rhoSqrReal, hybridXc, errStatus)
+      & densityMatrix, rhoPrim, work, rhoSqrReal, hybridXc, apiCallBack, errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3755,10 +3912,13 @@ contains
     !> Data for hybrid xc-functional calculation
     class(THybridXcFunc), intent(in), allocatable :: hybridXc
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
+    type(Tapicallback), intent(in), allocatable :: apiCallBack
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
-    integer :: iKS, iSpin
+    integer :: iKS, iK, iSpin
 
     rhoPrim(:,:) = 0.0_dp
 
@@ -3830,6 +3990,15 @@ contains
       end if
     #:endif
 
+      if (allocated(apiCallBack)) then
+        iK = parallelKS%localKS(1, iKS)
+      #:if WITH_SCALAPACK
+        call apiCallBack%invokeDM(iK, iSpin, work, denseDesc%blacsOrbSqr)
+      #:else
+        call apiCallBack%invokeDM(iK, iSpin, work)
+      #:endif
+      endif
+
       ! Store full density matrix for linear-response excited gradient evaluation
       ! (at this point equivalent to deltaRhoOut, but deltaRhoOut is later transformed into
       ! delta-density, therefore we store it separately for now)
@@ -3864,7 +4033,7 @@ contains
   !> Creates sparse density matrix from complex eigenvectors.
   subroutine getDensityFromCplxEigvecs(env, denseDesc, filling, kPoint, kWeight, neighbourList,&
       & nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, orb, parallelKS, tHelical,&
-      & species, coord, eigvecs, densityMatrix, rhoPrim, work, hybridXc, errStatus)
+      & species, coord, eigvecs, densityMatrix, rhoPrim, work, hybridXc, apiCallBack, errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3929,6 +4098,9 @@ contains
     !> Data for hybrid xc-functional calculation
     class(THybridXcFunc), intent(in), allocatable :: hybridXc
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
+    type(Tapicallback), intent(in), allocatable :: apiCallBack
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
@@ -3983,6 +4155,16 @@ contains
           densityMatrix%deltaRhoOutCplx(:,:, iKS) = work
         end if
       end if
+
+      if (allocated(apiCallBack)) then
+        iK = parallelKS%localKS(1, iKS)
+      #:if WITH_SCALAPACK
+        call apiCallBack%invokeDM(iK, iSpin, work, denseDesc%blacsOrbSqr)
+      #:else
+        call apiCallBack%invokeDM(iK, iSpin, work)
+      #:endif
+      endif
+
     end do
 
   #:if WITH_SCALAPACK
@@ -4004,7 +4186,7 @@ contains
   subroutine getDensityFromPauliEigvecs(env, denseDesc, tRealHS, tSpinOrbit, tDualSpinOrbit,&
       & tMulliken, kPoint, kWeight, filling, neighbourList, nNeighbourSK, orb, iSparseStart,&
       & img2CentCell, iCellVec, cellVec, species, parallelKS, eigvecs, work, dftbEnergy,&
-      & densityMatrix, rhoPrim, xi, orbitalL, iRhoPrim, errStatus)
+      & densityMatrix, rhoPrim, xi, orbitalL, iRhoPrim, apiCallBack, errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -4084,6 +4266,9 @@ contains
     !> Imaginary part of density matrix  if required
     real(dp), intent(inout), allocatable :: iRhoPrim(:,:)
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
+    type(Tapicallback), intent(in), allocatable :: apiCallBack
+
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
@@ -4121,9 +4306,20 @@ contains
       call densityMatrix%getDensityMatrix(work, eigvecs(:,:,iKS), filling(:,iK), errStatus)
       @:PROPAGATE_ERROR(errStatus)
     #:endif
+
       if (allocated(densityMatrix%deltaRhoOutCplx)) then
         densityMatrix%deltaRhoOutCplx(:,:,iKS) = 0.5_dp * work
       end if
+
+      if (allocated(apiCallBack)) then
+        iK = parallelKS%localKS(1, iKS)
+      #:if WITH_SCALAPACK
+        call apiCallBack%invokeDM(iK, 1, work, denseDesc%blacsOrbSqr)
+      #:else
+        call apiCallBack%invokeDM(iK, 1, work)
+      #:endif
+      endif
+
       if (tSpinOrbit .and. .not. tDualSpinOrbit) then
         call getOnsiteSpinOrbitEnergy(env, rVecTemp, work, denseDesc, xi, orb, species)
         dftbEnergy%atomLS = dftbEnergy%atomLS + kWeight(iK) * rVecTemp
@@ -4165,6 +4361,7 @@ contains
       end if
     #:endif
       call env%globalTimer%stopTimer(globalTimers%denseToSparse)
+
     end do
 
   #:if WITH_SCALAPACK
@@ -8111,7 +8308,7 @@ contains
   !> Creates (delta) density matrix for each microstate from real eigenvectors.
   subroutine getDensityMatrixL(env, denseDesc, neighbourList, nNeighbourSK, iSparseStart,&
       & img2CentCell, orb, species, coord, tPeriodic, tHelical, eigvecs, parallelKS, rhoPrim, work,&
-      & rhoSqrReal, q0, densityMatrix, hybridXc, reks, errStatus)
+      & rhoSqrReal, q0, densityMatrix, hybridXc, reks, apiCallBack, errStatus)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -8173,6 +8370,9 @@ contains
     !> Data type for REKS
     type(TReksCalc), intent(inout) :: reks
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(TAPICallback), intent(inout), allocatable :: apiCallBack
+
     !> Error status
     type(TStatus), intent(out) :: errStatus
 
@@ -8190,7 +8390,8 @@ contains
 
       call getDensityFromRealEigvecs(env, denseDesc, reks%fillingL(:,:,iL), neighbourList,&
           & nNeighbourSK, iSparseStart, img2CentCell, orb, species, coord, tPeriodic, tHelical,&
-          & eigvecs, parallelKS, densityMatrix, rhoPrim, work, rhoSqrReal, hybridXc, errStatus)
+          & eigvecs, parallelKS, densityMatrix, rhoPrim, work, rhoSqrReal, hybridXc, apiCallBack,&
+          & errStatus)
       @:PROPAGATE_ERROR(errStatus)
 
       if (reks%tForces) then