+++ src/dftbp/dftbplus/parser.F90
@@ -1251,6 +1251,7 @@ contains
     type(TStatus), intent(inout) :: errStatus
 
     type(string) :: buffer
+    type(fnode), pointer :: child
 
     call getNodeName(node, buffer)
     select case (char(buffer))
@@ -1271,6 +1272,15 @@ contains
     case default
       call detailedError(node, "Invalid Hamiltonian")
     end select
+
+  #:if WITH_API
+    call getChild(node, "ASI", child, requested=.false.)
+    if (associated(child)) then
+      ctrl%isASICallbackEnabled = .true.
+    else
+      ctrl%isASICallbackEnabled = .false.
+    end if
+  #:endif
 
   end subroutine readHamiltonian
 
@@ -2461,7 +2471,7 @@ contains
   end subroutine readSpinPolarisation
 
 
-  ! External field(s) and potential(s)
+  !> External field(s) and potential(s)
   subroutine readExternal(node, ctrl, geo)
 
     !> Relevant node in input tree